+++ example-compose-app/build.gradle
@@ -12,7 +12,7 @@ android {
     namespace 'com.hcaptcha.example.compose'
 
     defaultConfig {
-        minSdkVersion 23
+        minSdkVersion 21
         targetSdkVersion intProp("exampleTargetSdkVersion", 35)
         versionCode 1
         versionName "0.0.1"
